Bump k8s.io/kubernetes to v1.27.8, update tests
Previously we were on K8s v1.24 which is no longer supported in AKS.
Update to v1.27.8 so we're on an LTS version.
This also resolves some CVEs that were picked up by trivy,
although none of them affect ip-masq-agent.

Unfortunately, the test utilities we were using from k8s 1.24
changed significantly in k8s 1.25
https://github.com/kubernetes/kubernetes/commit/913f4bc0bab12534d91211703abff242cd57654f

Update the tests to use the new test utilities. This is a port of the work from
upstream ip-masq-agent:
https://github.com/kubernetes-sigs/ip-masq-agent/commit/8f9743a45c4c3814515403e2ba70cbde3454ca5d